chore: complete color utils
